+++ src/olmo_core/optim/adamw.py
@@ -28,6 +28,7 @@ def adamw_step(
     exp_avg_sq: torch.Tensor,
     step: torch.Tensor,
     step_factor: torch.Tensor,
+    step_increment_bugfix: bool = True,
 ):
     if p.grad is None:
         return
@@ -52,7 +53,8 @@ def adamw_step(
     update = -step_size * torch.div(exp_avg, denom)
     update.mul_(step_factor)
     p.add_(update)
-    step.add_(step_factor)
+    if step_increment_bugfix:
+        step.add_(step_factor)
 
 
 def foreach_adamw_step(
@@ -67,6 +69,7 @@ def foreach_adamw_step(
     eps: float,
     weight_decay: float,
     step_factor: torch.Tensor,
+    step_increment_bugfix: bool = True,
 ):
     """Perform a single AdamW update with multi-tensor (*foreach*) kernels."""
     if not params:
@@ -105,7 +108,8 @@ def foreach_adamw_step(
     updates = torch._foreach_div(exp_avgs, denoms)
     torch._foreach_mul_(updates, (-step_factor * step_sizes).unbind())
     torch._foreach_add_(params, updates)
-    torch._foreach_add_(steps, [step_factor] * len(steps))
+    if step_increment_bugfix:
+        torch._foreach_add_(steps, [step_factor] * len(steps))
 
 
 class SkipStepAdamW(SkipStepOptimizer):
@@ -124,6 +128,7 @@ class SkipStepAdamW(SkipStepOptimizer):
         sigma_factor: int = 6,
         dtype: Optional[Union[torch.dtype, DType]] = None,
         foreach: bool = False,
+        step_increment_bugfix: bool = True,
     ) -> None:
         assert lr > 0.0
         assert all([0.0 <= beta <= 1.0 for beta in betas])
@@ -138,6 +143,7 @@ class SkipStepAdamW(SkipStepOptimizer):
             dtype = dtype.as_pt()
         self.dtype = dtype
         self.foreach = foreach
+        self.stepfix = step_increment_bugfix
         self._step_skipped: Optional[torch.Tensor] = None
 
     @property
@@ -185,6 +191,7 @@ class SkipStepAdamW(SkipStepOptimizer):
                     exp_avg_sq=state["exp_avg_sq"],
                     step=state["step"],
                     step_factor=step_factor,
+                    step_increment_bugfix=self.stepfix,
                 )
 
     def _step_foreach(self, closure=None) -> None:
@@ -231,6 +238,7 @@ class SkipStepAdamW(SkipStepOptimizer):
                 eps=group["eps"],
                 weight_decay=group["weight_decay"],
                 step_factor=step_factor,
+                step_increment_bugfix=self.stepfix,
             )
             # grads_size = sum([g.numel() * g.element_size() for g in grads])/1024**3
             # exp_avgs_size = sum([ea.numel() * ea.element_size() for ea in exp_avgs])/1024**3
@@ -270,6 +278,15 @@ class SkipStepAdamWConfig(OptimConfig):
     """
     Whether to use multi-tensor (*foreach*) kernels for the AdamW update.
     Faster than the non-foreach version.
+    """
+
+    step_increment_bugfix: bool = True
+    """
+    Whether or not to fix the step-incrementing bug discovered in SkipStepAdamW.
+
+    If this flag is set to False, the step will not be incremented, which
+    gives the optimizer an effective lr that is 2.2x higher than the specified lr,
+    and no bias correction is applied.
     """
 
     rolling_interval_length: int = 128